fix custumor.wallet_money_amount type

+++ app/models/customer.py
@@ -8,7 +8,7 @@ class Customer(Base):
     user_id = Column(Integer, ForeignKey("user.id"), unique=True, nullable=False)
     subscription_model = Column(Enum("free", "plus", "premium", name="subscription_models"), nullable=False)
     subscription_end_time = Column(DateTime)
-    wallet_money_amount = Column(Numeric(10, 2), default=0.00, nullable=False)
+    wallet_money_amount = Column(Integer, default=0, nullable=False)
 
     # Relationships
     user = relationship("User", back_populates="customer")
